app/requirements.txt
psycopg2==2.8.4
Flask==1.1.1
Flask-Cors==3.0.8
python-dateutil